+++ module/src/commands/handlers/index.ts
@@ -94,7 +94,12 @@ export {
 } from '@/commands/handlers/scene';
 
 // World handlers
-export { getWorldInfoHandler, getCompendiumsHandler, getCompendiumHandler } from '@/commands/handlers/world';
+export {
+  getWorldInfoHandler,
+  getCompendiumsHandler,
+  getCompendiumHandler,
+  findInCompendiumHandler,
+} from '@/commands/handlers/world';
 
 // Table handlers
 export {