+++ module/src/commands/handlers/world/index.ts
@@ -1,3 +1,4 @@
 export { getWorldInfoHandler } from './GetWorldInfoHandler';
 export { getCompendiumsHandler } from './GetCompendiumsHandler';
 export { getCompendiumHandler } from './GetCompendiumHandler';
+export { findInCompendiumHandler } from './FindInCompendiumHandler';